moved fiat->crypto switching action to the amount preview text
